+++ src/attachments.ts
@@ -1,5 +1,12 @@
-import type { SVGandPNG } from './types';
+import type {
+  BarrelSlug,
+  GripSlug,
+  SVGandPNG,
+  SightSlug,
+  UnderBarrelSlug
+} from './types';
 import { getAssetURL, parseSvgString } from './utils';
+/** --- */
 import sightIcon from './assets/attachments/svgs/sight.svg';
 import barrelIcon from './assets/attachments/svgs/barrel.svg';
 import gripIcon from './assets/attachments/svgs/grip.svg';
@@ -83,28 +90,6 @@ export interface Sight {
   description: string;
   icon: SVGandPNG;
 }
-export type SightSlug =
-  | 'red_dot_a'
-  | 'red_dot_b'
-  | 'red_dot_c'
-  | 'red_dot_handgun'
-  | 'holo_a'
-  | 'holo_b'
-  | 'holo_c'
-  | 'holo_d'
-  | 'reflex_a'
-  | 'reflex_b'
-  | 'reflex_c'
-  | 'reflex_d'
-  | 'scope_1.5x'
-  | 'scope_2.0x'
-  | 'scope_2.5x_a'
-  | 'scope_2.5x_b'
-  | 'scope_3.0x_.44_mag'
-  | 'scope_3.0x'
-  | 'scope_4.0x'
-  | 'scope_5.0x'
-  | 'scope_12.0x';
 
 export const SIGHTS = [
   {
@@ -320,12 +305,6 @@ export interface Barrel {
   description: string;
   icon: SVGandPNG;
 }
-export type BarrelSlug =
-  | 'flash_hider'
-  | 'compensator'
-  | 'muzzle_brake'
-  | 'suppressor'
-  | 'extended_barrel';
 
 export const BARRELS = [
   {
@@ -384,7 +363,6 @@ export interface Grip {
   description: string;
   icon: SVGandPNG;
 }
-export type GripSlug = 'vertical_grip' | 'angled_grip';
 
 export const GRIPS = [
   {
@@ -415,7 +393,6 @@ export interface UnderBarrel {
   description: string;
   icon: SVGandPNG;
 }
-export type UnderBarrelSlug = 'laser';
 
 export const UNDER_BARRELS = [
   {
