+++ src/operators/index.ts
@@ -1,13 +1,13 @@
-import type { Operator, OperatorWeaponAttachments } from './types';
+import type { Season } from '../types';
 import { getNextPriceDropSeasons, getPrices } from '../utils';
 import { SEASONS } from '../seasons';
 import { type Gadget, GADGETS } from '../gadgets';
 import { type Weapon, WEAPONS } from '../weapons';
 import { BARRELS, GRIPS, SIGHTS, UNDER_BARRELS } from '../attachments';
+import type { Operator, OperatorWeaponAttachments } from './types';
 /** --- */
 import { recruitDefense } from './recruit_defense';
 import { recruitAttack } from './recruit_attack';
-import type { Season } from '../types';
 import { smoke } from './smoke';
 import { mute } from './mute';
 import { sledge } from './sledge';
